Refactor property and unit management; update repositories and services to include user ID for ownership validation

+++ BookingService.Shared.Infrastructure/Service/BaseRangeService.cs
@@ -1,3 +1,4 @@
+using BookingService.Shared.Infrastructure.Exceptions;
 using BookingService.Shared.Infrastructure.Repository;
 
 namespace BookingService.Shared.Infrastructure.Service;
@@ -11,8 +12,12 @@ public abstract class BaseRangeService<TModel, TCreateDto, TUpdateDto>(IBaseRang
     protected abstract TModel MapCreate(int id, TCreateDto dto);
     protected abstract TModel MapUpdate(int id, TUpdateDto dto);
     
-    public async Task AddRangeAsync(int id, List<TCreateDto> createDto)
+    public async Task AddRangeAsync(int id, int userId, List<TCreateDto> createDto)
     {
+        var parent = await repository.GetByOwnerIdAsync(userId);
+        if (parent.GetType().GetProperty("OwnerId")!.GetValue(parent)!.Equals(id))
+            throw new ForbidException("Parent owner ID does not match the user's ID.");
+        
         var modelList = createDto
             .Select(dto => MapCreate(id, dto))
             .ToList();
@@ -20,8 +25,12 @@ public abstract class BaseRangeService<TModel, TCreateDto, TUpdateDto>(IBaseRang
         await repository.AddRangeAsync(modelList);
     }
 
-    public async Task UpdateRangeAsync(int id, List<TUpdateDto> updateDto)
+    public async Task UpdateRangeAsync(int id, int userId, List<TUpdateDto> updateDto)
     {
+        var parent = await repository.GetByOwnerIdAsync(userId);
+        if (parent.GetType().GetProperty("OwnerId")!.GetValue(parent)!.Equals(id))
+            throw new ForbidException("Parent owner ID does not match the user's ID.");
+        
         var modelList = updateDto
             .Select(dto => MapUpdate(id, dto))
             .ToList();
@@ -29,8 +38,12 @@ public abstract class BaseRangeService<TModel, TCreateDto, TUpdateDto>(IBaseRang
         await repository.UpdateRangeAsync(modelList);
     }
     
-    public async Task DeleteRangeAsync(int id)
+    public async Task DeleteRangeAsync(int id, int userId)
     {
+        var parent = await repository.GetByOwnerIdAsync(userId);
+        if (parent.GetType().GetProperty("OwnerId")!.GetValue(parent)!.Equals(id))
+            throw new ForbidException("Parent owner ID does not match the user's ID.");
+        
         await repository.DeleteRangeAsync(id);
     }
 }
